fix subproc stdio
When in non-raw mode, subprocs now can read stdin, as expected.

When in raw mode, subprocs are always in a non-TTY and have no stdin.

To fix stdin in non-raw mode, we had to give up on process groups, because they
seem to break TTY detection for subprocs, and look for descendant procs by
ourselves, in platform-specific ways, falling back on `ps`.